backend/user_microservice/requirements.txt
flask
load_dotenv
pytest
